% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tseriesControl}
\alias{tseriesControl}
\title{Default Cross-validation control}
\usage{
tseriesControl(stepSize = 1, maxHorizon = 1, minObs = 12,
  fixedWindow = TRUE, summaryFunc = tsSummary, preProcess = FALSE,
  ppMethod = "guerrero")
}
\description{
Default Cross-validation control
}

